+++ src/main/java/com/amcamp/domain/image/application/ImageService.java
@@ -1,4 +1,4 @@
-package com.amcamp.domain.image.controller;
+package com.amcamp.domain.image.application;
 
 import com.amazonaws.HttpMethod;
 import com.amazonaws.services.s3.AmazonS3;
